Move footer and header, add header and footer option to Layout.js, add important pages
