+++ src/components/Layout.js
@@ -1,5 +1,14 @@
-const Layout = ({children}) => {
-  return <main>{children}</main>
+import Header from "./Header";
+import Footer from "./Footer";
+
+const Layout = ({children, header, footer, className}) => {
+  return (
+      <>
+        {header ? <Header /> : null}
+        <main className={className}>{children}</main>
+        {footer ? <Footer /> : null}
+      </>
+  )
 }
 
 export default Layout
