+++ src/pages/HomePage.js
@@ -1,28 +1,40 @@
 import Layout from "../components/Layout";
 import ToggleButton from "../components/ToggleButton";
 import {faChevronLeft, faChevronRight} from "@fortawesome/free-solid-svg-icons";
+
 const HomePage = () => {
   return (
-      <Layout>
+      <Layout header footer>
         <h1>Go Junho</h1>
-        <ToggleButton icon={faChevronLeft} />
-        <ToggleButton icon={faChevronRight} />
+        <ToggleButton icon={faChevronLeft}/>
+        <ToggleButton icon={faChevronRight}/>
         <p>Han Sooyoung</p>
         <div>Lorem ipsum</div>
         <div>Lorem ipsum</div>
-        <div>Lorem ipsum</div><div>Lorem ipsum</div><div>Lorem ipsum</div>
         <div>Lorem ipsum</div>
         <div>Lorem ipsum</div>
         <div>Lorem ipsum</div>
-        <div>Lorem ipsum</div><div>Lorem ipsum</div><div>Lorem ipsum</div>
         <div>Lorem ipsum</div>
-        <div>Lorem ipsum</div><div>Lorem ipsum</div><div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
         v
-        <div>Lorem ipsum</div><div>Lorem ipsum</div>
         <div>Lorem ipsum</div>
-        <div>Lorem ipsum</div><div>Lorem ipsum</div><div>Lorem ipsum</div><div>Lorem ipsum</div>
         <div>Lorem ipsum</div>
-        v<div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        <div>Lorem ipsum</div>
+        v
+        <div>Lorem ipsum</div>
         <div>Lorem ipsum</div>
         <div>Lorem ipsum</div>
         <div>Lorem ipsum</div>
