+++ src/pages/ListingPage.js
@@ -1,0 +1,9 @@
+import Layout from "../components/Layout";
+
+const ListingPage = () => {
+  return (
+      <Layout header footer></Layout>
+  )
+}
+
+export default ListingPage
