+++ src/pages/ProductDetailsPage.js
@@ -1,0 +1,9 @@
+import Layout from "../components/Layout";
+
+const ProductDetailsPage = () => {
+  return (
+      <Layout header footer></Layout>
+  )
+}
+
+export default ProductDetailsPage
