+++ src/pages/ProfilePage.js
@@ -1,0 +1,9 @@
+import Layout from "../components/Layout";
+
+const ProfilePage = () => {
+  return (
+      <Layout header footer></Layout>
+  )
+}
+
+export default ProfilePage
