+++ src/pages/auth/LoginPage.js
@@ -1,0 +1,9 @@
+import Layout from "../../components/Layout";
+
+const LoginPage = () => {
+  return (
+      <Layout></Layout>
+  )
+}
+
+export default LoginPage
