+++ src/pages/auth/RegisterPage.js
@@ -1,0 +1,9 @@
+import Layout from "../../components/Layout";
+
+const RegisterPage = () => {
+  return (
+      <Layout></Layout>
+  )
+}
+
+export default RegisterPage
